% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ru_count.R
\name{ru_count}
\alias{ru_count}
\title{Считает число терминов по каждой из групп по трем аспектам}
\usage{
ru_count(text)
}
\arguments{
\item{text}{Лемматизированный текст в виде строки, леммы в которой разделены одиночными пробелами.}
}
\value{
Список (list) количества терминов по отдельным темам в разрезе трех аспектов.
}
\description{
Эта функция подсчитывает количество вхождений по каждой из тематических групп
в разрезе аспектов деятельности компаний:
экологической,
социальной,
управленческой.
}
