% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ru_count_matches.R
\name{ru_count_matches}
\alias{ru_count_matches}
\title{Считает число вхождений паттерна в строку. Не нужно вызывать напрямую.}
\usage{
ru_count_matches(pat, vec)
}
\arguments{
\item{text}{Лемматизированный текст в виде строки, леммы в которой разделены одиночными пробелами.}
}
\value{
Список (list) количества терминов по отдельным темам в разрезе трех аспектов.
}
\description{
Эта функция используется фунцкией ru_count
}
